% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CodeWAS.R
\name{assertAnalysisSettings_CodeWAS}
\alias{assertAnalysisSettings_CodeWAS}
\title{Assert Analysis Settings for CodeWAS}
\usage{
assertAnalysisSettings_CodeWAS(analysisSettings)
}
\arguments{
\item{analysisSettings}{A list containing analysis settings. It must include the following elements:
\describe{
\item{cohortIdCases}{A numeric value representing the cohort ID for cases.}
\item{cohortIdControls}{A numeric value representing the cohort ID for controls.}
\item{analysisIds}{A numeric vector of analysis IDs.}
\item{covariatesIds}{A numeric vector of covariate IDs (optional).}
\item{minCellCount}{A numeric value representing the minimum cell count, must be 0 or higher.}
\item{chunksSizeNOutcomes}{A numeric value representing the chunk size for outcomes (optional).}
\item{cores}{A numeric value representing the number of cores to use for parallel processing.}
}}
}
\value{
Returns the validated \code{analysisSettings} list.
}
\description{
Validates the \code{analysisSettings} list to ensure it contains the required elements (\code{cohortIdCases}, \code{cohortIdControls}, \code{analysisIds}, \code{covariatesIds}, \code{minCellCount}, \code{chunksSizeNOutcomes}, \code{cores}) with correct types and values. This function is specifically designed for checking settings related to CodeWAS analysis.
}
