% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CohortDemographics.R
\name{assertAnalysisSettings_CohortDemographics}
\alias{assertAnalysisSettings_CohortDemographics}
\title{Assert Analysis Settings for Cohort Demographics Overlaps}
\usage{
assertAnalysisSettings_CohortDemographics(analysisSettings)
}
\arguments{
\item{analysisSettings}{A list containing analysis settings. It must include the following elements:
\describe{
\item{cohortIds}{A numeric vector of cohort IDs.}
\item{referenceYears}{A character vector specifying the reference years, must be one of \code{cohort_start_date}, \code{cohort_end_date}, or \code{birth_datetime}.}
\item{groupBy}{A character vector indicating the demographic groups for analysis, must be one of \code{calendarYear}, \code{ageGroup}, or \code{gender}.}
\item{minCellCount}{A numeric value representing the minimum cell count, must be 0 or higher.}
}}
}
\value{
Returns \code{TRUE} if all settings are valid; otherwise, throws an error.
}
\description{
Validates the \code{analysisSettings} list to ensure it contains the required elements (\code{cohortIds}, \code{referenceYears}, \code{groupBy}, \code{minCellCount}) with correct types and values. This function is specifically designed for checking settings related to cohort demographics analysis.
}
