% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CohortOverlaps.R
\name{assertAnalysisSettings_CohortOverlaps}
\alias{assertAnalysisSettings_CohortOverlaps}
\title{Assert Analysis Settings for Cohort Overlaps}
\usage{
assertAnalysisSettings_CohortOverlaps(analysisSettings)
}
\arguments{
\item{analysisSettings}{A list containing analysis settings, including \code{cohortIds} and \code{minCellCount}.}
}
\value{
TRUE if the settings are valid; otherwise, an error is thrown.
}
\description{
This function checks that the \code{analysisSettings} list contains the required elements and that they are of the correct types.
}
