% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_GWAS.R
\name{assertAnalysisSettings_GWAS}
\alias{assertAnalysisSettings_GWAS}
\title{Assert Analysis Settings for GWAS}
\usage{
assertAnalysisSettings_GWAS(analysisSettings)
}
\arguments{
\item{analysisSettings}{A list containing analysis settings. It must include the following elements:
\describe{
\item{casesFinngenids}{A vector of strings containing cases FinnGen IDs.}
\item{controlsFinngenids}{A vector of strings containing controls FinnGen IDs.}
\item{analysisType}{String containing analysis type (additive, recessive, or dominant).}
\item{phenotype}{String containing phenotype name.}
\item{description}{String containing description for the custom GWAS analysis.}

}}
}
\value{
Returns the validated \code{analysisSettings} list.
}
\description{
Validates the \code{analysisSettings} list to ensure it contains the required elements
}
