% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CodeWAS.R
\name{checkResults_CodeWAS}
\alias{checkResults_CodeWAS}
\title{Check CodeWAS Results}
\usage{
checkResults_CodeWAS(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{A string representing the path to the DuckDB database file containing the results.}
}
\value{
TRUE if the results are valid; otherwise, a vector of error messages.
}
\description{
This function checks the integrity and correctness of the exported CodeWAS results in a DuckDB database.
}
