% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CohortDemographics.R
\name{checkResults_CohortDemographics}
\alias{checkResults_CohortDemographics}
\title{Check Cohort Overlaps Results}
\usage{
checkResults_CohortDemographics(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{A string representing the path to the DuckDB database file containing the results.}
}
\value{
TRUE if the results are valid; otherwise, a vector of error messages.
}
\description{
This function checks the integrity and correctness of the exported cohort overlaps results in a DuckDB database.
}
