% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_PhenotypeScoring.R
\name{checkResults_PhenotypeScoring}
\alias{checkResults_PhenotypeScoring}
\title{Check PhenotypeScoring Results}
\usage{
checkResults_PhenotypeScoring(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{A string representing the path to the DuckDB database file containing the results.}
}
\value{
TRUE if the results are valid; otherwise, a vector of error messages.
}
\description{
This function checks the integrity and correctness of the exported PhenotypeScoring results in a DuckDB database.
}
