% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_timeCodeWAS.R
\name{checkResults_timeCodeWAS}
\alias{checkResults_timeCodeWAS}
\title{Check TimeCodeWAS Results}
\usage{
checkResults_timeCodeWAS(pathToResultsDatabase)
}
\arguments{
\item{pathToResultsDatabase}{A string representing the path to the DuckDB database file containing the results.}
}
\value{
TRUE if the results are valid; otherwise, a vector of error messages.
}
\description{
This function checks the integrity and correctness of the exported timeCodeWAS results in a DuckDB database.
}
