% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tibbleStatisticalTests.R
\name{.addTestToCodeCounts}
\alias{.addTestToCodeCounts}
\title{.addFisherTestToCodeCounts}
\usage{
.addTestToCodeCounts(tibbleWithCodeCounts)
}
\arguments{
\item{tibbleWithCodeCounts}{table obtained from \code{getCodeCounts}.}
}
\value{
imputed table with append columns :
\itemize{
\item p : for the p-value
\item OR; for the odds ratio
}
}
\description{
performs an Fisher test in each row of the case_controls_counts table
}
