% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tibbleStatisticalTests.R
\name{.addTestTotibbleWithValueSummary}
\alias{.addTestTotibbleWithValueSummary}
\title{.addTestTotibbleWithValueSummary}
\usage{
.addTestTotibbleWithValueSummary(tibbleWithValueSummary)
}
\arguments{
\item{tibbleWithValueSummary}{table containing summary statistics for cases and controls.}
}
\value{
input table with appended columns:
\itemize{
\item continuousPValue: the p-value of the test
\item continuousStandardError: the standard error of the test
\item continuousTest: the type of test performed
}
}
\description{
performs a continuous test in each row of the tibbleWithValueSummary table
}
