% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tibbleStatisticalTests.R
\name{.binaryTest}
\alias{.binaryTest}
\title{.binaryTest}
\usage{
.binaryTest(a, b, c, d, fiserLimit = 10)
}
\arguments{
\item{a}{number of cases with the condition}

\item{b}{number of cases without the condition}

\item{c}{number of controls with the condition}

\item{d}{number of controls without the condition}

\item{fiserLimit}{threshold to decide between Fisher and Chi-square test}
}
\value{
a list with:
\itemize{
\item \code{countsPValue}: the p-value of the test
\item \code{countsOddsRatio}: the odds ratio
\item \code{countsTest}: the type of test performed
}
}
\description{
performs a Fisher or Chi-square test based on the input counts
}
