% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.calculateTotalScores}
\alias{.calculateTotalScores}
\title{Calculate Total Scores}
\usage{
.calculateTotalScores(groupOfCovariatesObject, formula)
}
\arguments{
\item{groupOfCovariatesObject}{A list containing the group of covariates object}

\item{formula}{A string containing the formula to calculate the total scores}
}
\value{
A list containing the updated group of covariates object
}
\description{
Calculates the total scores for each person in the groupOfCovariatesObject
}
