% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{.checkDatabase}
\alias{.checkDatabase}
\title{Check Database Schema}
\usage{
.checkDatabase(pathToResultsDatabase, expectedSchemas)
}
\arguments{
\item{pathToResultsDatabase}{A string representing the path to the DuckDB database file.}

\item{expectedSchemas}{A named list of tibbles, where each tibble represents the expected schema of a database table with columns \code{name} and \code{type}.}
}
\value{
Returns \code{TRUE} if all scheme match the expected scheme. Otherwise, it returns a list of error messages indicating the discrepancies.
}
\description{
This function checks the schema of a DuckDB database against expected schemas.
}
