% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_GWAS.R
\name{.configGWAS}
\alias{.configGWAS}
\title{Configure GWAS Sandbox API Connection}
\usage{
.configGWAS()
}
\value{
An object representing the connection to the sandbox API.
}
\description{
This function configures and retrieves a connection to the internal GWAS sandbox API. It refreshes the authorization token
and handles errors related to SSL verification and token refreshment.
}
\details{
The function refreshes the token by making an API call, updates the token in the environment variable \code{SANDBOX_TOKEN},
and returns the API connection object.
}
