% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tibbleStatisticalTests.R
\name{.continuousTest}
\alias{.continuousTest}
\title{.continuousTest}
\usage{
.continuousTest(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1}{the sample mean of the first group}

\item{m2}{the sample mean of the second group}

\item{s1}{the sample standard deviation of the first group}

\item{s2}{the sample standard deviation of the second group}

\item{n1}{the sample size of the first group}

\item{n2}{the sample size of the second group}

\item{m0}{the null value for the difference in means to be tested for. Default is 0.}

\item{equal.variance}{whether or not to assume equal variance. Default is FALSE.}
}
\value{
a list with:
\itemize{
\item continuousPValue: the p-value of the test
\item continuousOddsRatio: odds ration calculated as meanValueCases/meanValueControls
\item continuousTest: the type of test performed
}
}
\description{
performs a Welch Two Sample t-test or a t-test with equal variance based on the input parameters
}
