% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysisSettings_GWAS.R
\name{.format_str}
\alias{.format_str}
\title{Format String by Removing Special Characters}
\usage{
.format_str(x)
}
\arguments{
\item{x}{A character string to be formatted.}
}
\value{
A character string in uppercase with special characters and spaces removed.
}
\description{
This function takes a string and formats it by converting it to uppercase and removing all
punctuation, non-alphanumeric characters, and spaces.
}
