% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.getAllCovariatesTibble}
\alias{.getAllCovariatesTibble}
\title{Get All Covariates Tibble}
\usage{
.getAllCovariatesTibble(analysisResults)
}
\arguments{
\item{analysisResults}{A database connection containing analysis results tables}
}
\value{
A tibble containing all covariates data with reference information
}
\description{
Retrieves all covariates data from analysis results and joins with reference tables
}
