% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CohortDemographics.R
\name{.getCohortDemographicsCounts}
\alias{.getCohortDemographicsCounts}
\title{Get Cohort Demographics Counts}
\usage{
.getCohortDemographicsCounts(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = c(),
  referenceYears = c("cohort_start_date", "cohort_end_date", "birth_datetime")
)
}
\arguments{
\item{connectionDetails}{Database connection details, created using \code{DatabaseConnector::createConnectionDetails()}. Either this or \code{connection} must be provided.}

\item{connection}{An existing database connection from \code{DatabaseConnector::connect()}. If \code{NULL}, a connection will be established using \code{connectionDetails}.}

\item{cdmDatabaseSchema}{The schema of the CDM database where the patient data resides.}

\item{vocabularyDatabaseSchema}{(Optional) The schema of the vocabulary tables. Defaults to the \code{cdmDatabaseSchema}.}

\item{cohortDatabaseSchema}{The schema of the cohort tables where cohorts are stored.}

\item{cohortTable}{The name of the cohort table. Default is "cohort".}

\item{cohortIds}{A numeric vector of cohort IDs to filter on. If \code{NULL} or empty, all cohorts are included.}

\item{referenceYears}{A character vector specifying the reference year(s) to use for demographic counts. Valid options are "cohort_start_date", "cohort_end_date", and "birth_datetime". Default is all three.}
}
\value{
A tibble containing demographic counts per cohort, grouped by age group and gender for each reference year. The columns include \code{referenceYear}, \code{cohortId}, \code{calendarYear}, \code{ageGroup}, \code{gender}, and \code{count}.
}
\description{
This function retrieves demographic counts (age and gender) for cohorts based on a reference year (cohort start date, cohort end date, or birth date). It queries the cohort table in the specified database and provides the demographic distribution by age group and gender.
}
