% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\docType{data}
\name{.numericRangeFilter}
\alias{.numericRangeFilter}
\title{Numeric Range Filter}
\format{
An object of class \code{JS_EVAL} of length 1.
}
\usage{
.numericRangeFilter
}
\value{
Returns a list of rows that match the filter.
}
\description{
Internal JS function to filter numeric values in reactable columns.
Supports ranges (e.g., -2--0.5, 1-4), comparison operators (>=, <=, >, <, ==), and bare numbers.
}
\keyword{datasets}
