% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CohortDemographics.R
\name{execute_CohortDemographics}
\alias{execute_CohortDemographics}
\title{execute_CohortDemographics}
\usage{
execute_CohortDemographics(exportFolder, cohortTableHandler, analysisSettings)
}
\arguments{
\item{exportFolder}{A string representing the path to the folder where the results will be exported.}

\item{cohortTableHandler}{An R6 object of class \code{CohortTableHandler} containing information about the cohort tables.}

\item{analysisSettings}{A list containing analysis settings, including \code{cohortIds} and \code{minCellCount}.}
}
\value{
A string representing the path to the exported results database.
}
\description{
This function calculates cohort overlaps based on the provided cohort table and analysis settings, and exports the results to a DuckDB database.
}
