% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysisSettings_CohortDemographics.R
\name{mod_analysisSettings_cohortDemographics_ui}
\alias{mod_analysisSettings_cohortDemographics_ui}
\title{Cohort Demographics Analysis Settings UI}
\usage{
mod_analysisSettings_cohortDemographics_ui(id)
}
\arguments{
\item{id}{A string representing the module's namespace.}
}
\value{
A Shiny UI element that can be included in a Shiny app.
}
\description{
UI module for configuring the settings for cohort overlaps analysis. This module allows users to select cohorts and set the minimum cell count for the analysis.
}
