% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysisSettings_GWAS.R
\name{mod_analysisSettings_GWAS_server}
\alias{mod_analysisSettings_GWAS_server}
\title{Analysis Settings for GWAS Module Server}
\usage{
mod_analysisSettings_GWAS_server(id, r_connectionHandler)
}
\arguments{
\item{id}{A string representing the module's ID.}

\item{r_connectionHandler}{A reactive object containing database and cohort connection handlers.}
}
\value{
A reactive expression that provides the analysis settings for the GWAS.
}
\description{
A Shiny module server function that sets up the analysis settings for a GWAS (Genome-Wide Association Study).
It manages the user inputs for selecting case and control cohorts, setting a phenotype name, and description,
and provides reactive outputs for the analysis settings.
}
