% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_formTimeWindows.R
\name{mod_fct_formTimeWindows_server}
\alias{mod_fct_formTimeWindows_server}
\title{Form Time Windows Server Module}
\usage{
mod_fct_formTimeWindows_server(id, session)
}
\arguments{
\item{id}{Module ID}

\item{session}{Shiny session object}
}
\value{
A reactive list containing the start and end days of the time windows
}
\description{
This module provides the server logic for forming time windows.
}
