% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_formTimeWindows.R
\name{mod_fct_formTimeWindows_ui}
\alias{mod_fct_formTimeWindows_ui}
\title{Form Time Windows UI Module}
\usage{
mod_fct_formTimeWindows_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A Shiny UI element
}
\description{
This module provides the UI for forming time windows.
}
