% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_CohortDemographics.R
\name{mod_resultsVisualisation_CohortsDemographics_ui}
\alias{mod_resultsVisualisation_CohortsDemographics_ui}
\title{Cohort Demographics Visualization UI}
\usage{
mod_resultsVisualisation_CohortsDemographics_ui(id)
}
\arguments{
\item{id}{A string representing the module's namespace.}
}
\value{
A Shiny UI element that can be included in a Shiny app.
}
\description{
UI module for visualizing cohort demographics. This module provides controls to customize the appearance of the plot and options to download the plot and data.
}
