% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation.R
\name{mod_resultsVisualisation_ui}
\alias{mod_resultsVisualisation_ui}
\title{Results Visualization Dashboard UI}
\usage{
mod_resultsVisualisation_ui(
  id,
  resultsVisualisationModuleUi,
  pathToAboutMd,
  title,
  logshref
)
}
\arguments{
\item{id}{A string representing the module's namespace.}

\item{resultsVisualisationModuleUi}{A UI function representing the specific visualization module to include in the main panel.}

\item{pathToAboutMd}{A string representing the path to a markdown file with information about the application.}

\item{title}{A string representing the title of the main tab in the sidebar.}

\item{logshref}{A string representing the URL to the logs page.}
}
\value{
A Shiny UI element that can be included in a Shiny app.
}
\description{
UI module for creating a Shiny dashboard that visualizes results, including cohort overlaps and cohort definitions. This module provides a sidebar with navigation options and a main panel for displaying the content.
}
