% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.appendCovariateGroup}
\alias{.appendCovariateGroup}
\title{Append Covariate Group}
\usage{
.appendCovariateGroup(analysisResults, covariateIds, groupOfCovariatesObject)
}
\arguments{
\item{analysisResults}{A database connection containing analysis results tables}

\item{covariateIds}{A vector of covariate ids}

\item{groupOfCovariatesObject}{A list containing the group of covariates object}
}
\value{
A list containing the updated group of covariates object
}
\description{
Appends a new group of covariates to the groupOfCovariatesObject
}
