% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.appendCovariateGroup}
\alias{.appendCovariateGroup}
\title{Append Covariate Group}
\usage{
.appendCovariateGroup(
  analysisResults,
  covariateIds,
  groupedCovariatesTibble,
  groupedCovariatesPerPersonTibble
)
}
\arguments{
\item{analysisResults}{A database connection containing analysis results tables}

\item{covariateIds}{A vector of covariate ids}

\item{groupedCovariatesTibble}{A tibble containing the grouped covariates}

\item{groupedCovariatesPerPersonTibble}{A tibble containing the grouped covariates per person}
}
\value{
A list containing the updated group of covariates object
}
\description{
Appends a new group of covariates to the groupOfCovariatesObject
}
