% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.calculateTotalScores}
\alias{.calculateTotalScores}
\title{Calculate Total Scores}
\usage{
.calculateTotalScores(groupedCovariatesPerPersonTibble, formula)
}
\arguments{
\item{groupedCovariatesPerPersonTibble}{A tibble containing the grouped covariates per person}

\item{formula}{A string containing the formula to calculate the total scores}
}
\value{
A tibble containing the updated \code{groupedCovariatesPerPersonTibble}, with a new column \code{totalScore} and \code{totalScoreBin}
}
\description{
Calculates the total scores for each person in the \code{groupedCovariatesPerPersonTibble} given a formula
}
