% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_resultsVisualisation_PhenotypeScoring.R
\name{.renderCovariatesDistribution}
\alias{.renderCovariatesDistribution}
\title{Render Covariates Distribution}
\usage{
.renderCovariatesDistribution(covariatesDistribution, metric = "count")
}
\arguments{
\item{covariatesDistribution}{A tibble containing the covariates distribution}
}
\value{
A plot of the covariates distribution
}
\description{
Renders a plot of the covariates distribution
}
